feat/fix: Masse tilføjelser og ændringer

+++ src/main/java/dk/zealand/hw_deckforge/presentation/controllers/CardController.java
@@ -19,47 +19,58 @@ public class CardController {
     }
 
     @GetMapping
-    public String getAllCards(Model model, HttpSession session) {
-        model.addAttribute("cards",cardService.getAll());
+    public String getAllCards(Model model) {
+        model.addAttribute("cards", cardService.getAll());
         return "cards/card-list";
+    }
+
+    @GetMapping("/{id}")
+    public String getCard(@PathVariable int id, Model model) {
+        model.addAttribute("card", cardService.getById(id));
+        return "cards/card-detail";
     }
 
     @GetMapping("/create")
     public String showCreateForm(Model model, HttpSession session) {
-        model.addAttribute("card", new Card(0, "", null, "",
-                null, null, null, null));
+        if (!AuthHelper.isAdmin(session)) return "redirect:/access-denied";
+        model.addAttribute("card", new Card());
         return "cards/add-card";
     }
 
     @PostMapping("/create")
     public String create(@ModelAttribute Card card, HttpSession session) {
+        if (!AuthHelper.isAdmin(session)) return "redirect:/access-denied";
         cardService.create(card);
         return "redirect:/cards";
     }
 
     @GetMapping("/{id}/edit")
     public String showEditForm(@PathVariable int id, Model model, HttpSession session) {
+        if (!AuthHelper.isAdmin(session)) return "redirect:/access-denied";
         model.addAttribute("card", cardService.getById(id));
         return "cards/edit-card";
     }
 
     @PostMapping("/{id}/edit")
     public String update(@PathVariable int id, @ModelAttribute Card card, HttpSession session) {
+        if (!AuthHelper.isAdmin(session)) return "redirect:/access-denied";
         card.setId(id);
         cardService.update(card);
         return "redirect:/cards";
     }
 
-
     @GetMapping("/{id}/delete")
     public String showDeleteConfirm(@PathVariable int id, Model model, HttpSession session) {
-        Card card = cardService.getById(id);
-        model.addAttribute("card",card);
-        return "delete-confirm";
+        if (!AuthHelper.isAdmin(session)) return "redirect:/access-denied";
+        model.addAttribute("card", cardService.getById(id));
+        model.addAttribute("tilbage", "/cards");
+        return "cards/delete-confirm";
     }
 
     @PostMapping("/{id}/delete")
     public String delete(@PathVariable int id, HttpSession session) {
+        if (!AuthHelper.isAdmin(session)) return "redirect:/access-denied";
         cardService.delete(id);
-        return "redirect:/cards"; }
+        return "redirect:/cards";
+    }
 }
